# server/api/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "BatchingService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/query/batch": {
      "post": {
        "operationId": "BatchingService_QueryBatch",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/v1QueryBatchResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of v1QueryBatchResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1QueryBatchRequest"
            }
          }
        ],
        "tags": [
          "BatchingService"
        ]
      }
    },
    "/v1/query/one/{name}": {
      "post": {
        "operationId": "BatchingService_QueryOne",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1QueryOneResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "argOneZero": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "BatchingService"
        ]
      }
    },
    "/v1/query/two/{name}": {
      "post": {
        "operationId": "BatchingService_QueryTwo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1QueryTwoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "argTwoZero": {
                  "type": "integer",
                  "format": "int32"
                },
                "argTwoOne": {
                  "type": "string"
                },
                "argTwoTwo": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "tags": [
          "BatchingService"
        ]
      }
    },
    "/v1/query/zero/{name}": {
      "post": {
        "operationId": "BatchingService_QueryZero",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1QueryZeroResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "argZeroZero": {
                  "type": "string"
                },
                "argZeroOne": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        ],
        "tags": [
          "BatchingService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1QueryBatchRequest": {
      "type": "object",
      "properties": {
        "queries": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1QueryRequest"
          }
        }
      }
    },
    "v1QueryBatchResponse": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/v1QueryResponse"
        }
      }
    },
    "v1QueryOneRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "argOneZero": {
          "type": "string"
        }
      }
    },
    "v1QueryOneResponse": {
      "type": "object",
      "properties": {
        "respOneZero": {
          "type": "string"
        },
        "respOneOne": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1QueryRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "$ref": "#/definitions/v1QueryType"
        },
        "zeroRequest": {
          "$ref": "#/definitions/v1QueryZeroRequest"
        },
        "oneRequest": {
          "$ref": "#/definitions/v1QueryOneRequest"
        },
        "twoRequest": {
          "$ref": "#/definitions/v1QueryTwoRequest"
        }
      }
    },
    "v1QueryResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "$ref": "#/definitions/v1QueryType"
        },
        "zeroResponse": {
          "$ref": "#/definitions/v1QueryZeroResponse"
        },
        "oneResponse": {
          "$ref": "#/definitions/v1QueryOneResponse"
        },
        "twoResponse": {
          "$ref": "#/definitions/v1QueryTwoResponse"
        },
        "error": {
          "type": "string"
        }
      }
    },
    "v1QueryTwoRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "argTwoZero": {
          "type": "integer",
          "format": "int32"
        },
        "argTwoOne": {
          "type": "string"
        },
        "argTwoTwo": {
          "type": "boolean"
        }
      }
    },
    "v1QueryTwoResponse": {
      "type": "object",
      "properties": {
        "respTwoZero": {
          "type": "string"
        }
      }
    },
    "v1QueryType": {
      "type": "string",
      "enum": [
        "Zero",
        "One",
        "Two"
      ],
      "default": "Zero"
    },
    "v1QueryZeroRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "argZeroZero": {
          "type": "string"
        },
        "argZeroOne": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1QueryZeroResponse": {
      "type": "object",
      "properties": {
        "respZeroZero": {
          "type": "string"
        }
      }
    }
  }
}